langchain_project/requirements.txt
langchain
langchain-openai
openai
faiss-cpu
python-dotenv
